Skipping test if server is too old for SimpleAggregateFunction
